Autopost/src/requirements.txt
PyYAML~=5.3.1
GitPython~=3.1.11
google-cloud-storage~=1.30.0
beautifulsoup4~=4.9.3
google-cloud-secret-manager